# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Elevator Control System API
  description: |
    A RESTful API for managing elevator systems in buildings. This API allows you to:
    - Request elevators to move between floors
    - Create new elevators in the system
    - Monitor system health and metrics
    - Get real-time status updates via WebSocket
  version: "1.0.0"
  contact:
    name: Elevator System Support
    email: support@elevator-system.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:6660
    description: Local development server
  - url: https://elevator-api.example.com
    description: Production server
security:
  - {}  # No authentication required for this version
paths:
...
  /ws/status:
    get:
      summary: WebSocket status updates
      description: |
        WebSocket endpoint for real-time elevator status updates.
        Connect to receive periodic status updates about all elevators.
      operationId: connectWebSocket
      tags:
        - Real-time Updates
      responses:
        '101':
          description: WebSocket connection established
        '400':
          description: Bad WebSocket request
...